refactor: chat 사용자 기능 개선
