+++ packages/gdsc-dju-admin/src/components/organisms/modal/styled.ts
